fix(justfile): make `just build` actually compile everything; add asan/tsan modes
`just build` was silently a no-op: the `find` filter in `_build-inner`
excluded every directory containing real Mojo code (shared/, examples/,
benchmarks/, papers/, tests/) and stripped __init__.mojo / model.mojo /
test_*.mojo / repro_*.mojo, leaving exactly one file —
`scripts/verify_installation.mojo` — to actually compile. The recipe then
printed "✅ Build successful" regardless.

Changes
-------

`_build-inner` rewritten to do two clear phases:

  Phase A — package the shared library (always runs):
    `mojo package --Werror -I . shared -o build/$MODE/ProjectOdyssey-shared.mojopkg`
    (mojo package doesn't accept -g / -O / --sanitize flags, so STRICT only.)

  Phase B — AOT-compile every executable with main():
    find examples/, benchmarks/, papers/*/examples/ + grep for `^fn main` /
    `^def main`. Each survivor compiles to build/$MODE/<flattened_name> with
    -Xlinker -lm (the same fix `_test-group-asan-inner` already uses for
    transitive libm symbols — supersedes the older "exclude examples/" recipe
    documented in skill `exclude-build-targets-from-linker-errors`).

  Output: 46 binaries + 1 .mojopkg per mode (vs. 1 binary before).

Werror is now real: --Werror is applied to every `mojo build` invocation
and the loop hard-fails on the first warning. No FAIL_ON_ERROR=0 escape
hatch.

New modes:
  - `just build asan` / `just build tsan` — package shared/ under the
    sanitizer flags; skip Phase B (mojo compiler under sanitizers needs
    3-6 GB per invocation, so a 46-file sweep OOMs hosts with <32 GB).
  - `just build-asan`, `just build-tsan` — convenience aliases.
  - `just build-all` extended to include asan + tsan.
  - tsan adds `-j1` (modular/modular#6413 workaround).

New test recipes:
  - `just test-mojo-asan` / `just test-mojo-tsan` — compile + run every
    test under the chosen sanitizer with -Werror. Exercises sanitizers
    where they actually pay off (runtime race / leak detection on the
    test suite, not AOT validation of example binaries).

Pre-existing bug fixed: `_test-mojo-inner` and the new sanitized variant
walked `tests/**/*.mojo` without `shopt -s globstar`. That glob only
matched one level deep, silently skipping ~85% of the test suite.
Replaced with a `find … | sort` + `mapfile` pattern that visits all 298
test files.

podman-up made clone-aware: if a container already exists with a
workspace mount pointing at a different clone (e.g. ~/ProjectOdyssey vs
~/Projects/ProjectOdyssey), it's torn down so `compose up` creates a
fresh container bound to the current $PWD. Avoids running stale code via
a leftover container from another working copy.

docker-compose.yml: bumped default mem_limit from 12g → 14g (leaves 1 GB
headroom on the typical 15 GB workstation) and documented the
ODYSSEY_MEM_LIMIT=24g override for sanitizer sweeps.

Werror cleanup in samples/benchmarks/papers (surfaced by the now-working
build):
  - Deprecated implicit stdlib imports (`from time import …` →
    `from std.time import …`) in examples/mojo_patterns/simd_example.mojo
    and four benchmarks/ files.
  - Deprecated Int→UInt8 implicit conversions in
    benchmarks/scripts/compare_results.mojo (`bytes[i] == ord("X")` →
    `bytes[i] == UInt8(ord("X"))`).
  - Docstring `Parameters:` / `Args:` slot mixup in
    benchmarks/scripts/run_benchmarks.mojo.
  - Unused loop variable in examples/mobilenetv1_cifar10/inference.mojo.
  - UnsafePointer-as-Bool deprecation in
    examples/mojo_patterns/trait_example.mojo.
  - tests/shared/test_imports.mojo: `from shared import SGD` was
    ambiguous (two SGD struct declarations: shared/training/__init__.mojo
    and shared/autograd/optimizers.mojo). Imported via shared.training
    only and documented the conflict; the duplicate-SGD library issue
    needs a separate fix.

Verification
------------
  - just build debug  →  46 binaries + 1 mojopkg
  - just build release  →  same
  - just build asan  →  shared mojopkg only (skips Phase B)
  - just build tsan  →  shared mojopkg only (skips Phase B)
  - just build-all  →  all five modes succeed
  - just test-mojo  →  298 / 298 pass under --Werror (0 compile errors)

Out of scope (follow-up issues recommended)
-------------------------------------------
ASAN test sweep ran 298/298 with 0 compile errors but exposed 46 tests
with memory leaks across shared/ (pooled_alloc, AnyTensor.__init__,
maxpool2d, tensor_creation::{ones,arange}). These are real library
leaks, not artifacts of the recipe.

TSAN test sweep runs but every binary aborts at startup with
`FATAL: ThreadSanitizer: unexpected memory mapping` /
`tcmalloc MmapAligned() failed` — a known TSAN/tcmalloc incompatibility
in the Mojo runtime, not a code race. The recipe is in place; getting it
green needs Modular-side runtime work.

+++ benchmarks/scripts/compare_results.mojo
@@ -18,7 +18,7 @@ Output:
 """
 
 from std.sys import argv
-from python import Python
+from std.python import Python
 
 
 # ============================================================================
@@ -123,37 +123,43 @@ def atof(s: String) -> Float64:
     var bytes = s.as_bytes()
 
     # Handle negative numbers
-    if len(bytes) > 0 and bytes[0] == ord("-"):
+    if len(bytes) > 0 and bytes[0] == UInt8(ord("-")):
         sign = -1.0
         i = 1
 
     # Parse integer part
     while (
         i < len(bytes)
-        and bytes[i] != ord(".")
-        and bytes[i] != ord("e")
-        and bytes[i] != ord("E")
+        and bytes[i] != UInt8(ord("."))
+        and bytes[i] != UInt8(ord("e"))
+        and bytes[i] != UInt8(ord("E"))
     ):
         result = result * 10.0 + Float64(Int(bytes[i]) - ord("0"))
         i += 1
 
     # Parse decimal part
-    if i < len(bytes) and bytes[i] == ord("."):
+    if i < len(bytes) and bytes[i] == UInt8(ord(".")):
         i += 1
         var decimal_place: Float64 = 0.1
-        while i < len(bytes) and bytes[i] != ord("e") and bytes[i] != ord("E"):
+        while (
+            i < len(bytes)
+            and bytes[i] != UInt8(ord("e"))
+            and bytes[i] != UInt8(ord("E"))
+        ):
             result = result + Float64(Int(bytes[i]) - ord("0")) * decimal_place
             decimal_place = decimal_place * 0.1
             i += 1
 
     # Parse exponent (scientific notation)
-    if i < len(bytes) and (bytes[i] == ord("e") or bytes[i] == ord("E")):
+    if i < len(bytes) and (
+        bytes[i] == UInt8(ord("e")) or bytes[i] == UInt8(ord("E"))
+    ):
         i += 1
         var exp_sign = 1
-        if i < len(bytes) and bytes[i] == ord("-"):
+        if i < len(bytes) and bytes[i] == UInt8(ord("-")):
             exp_sign = -1
             i += 1
-        elif i < len(bytes) and bytes[i] == ord("+"):
+        elif i < len(bytes) and bytes[i] == UInt8(ord("+")):
             i += 1
 
         var exponent = 0
@@ -327,13 +333,13 @@ def atoi(s: String) -> Int:
     var s_bytes = s.as_bytes()
 
     # Handle negative numbers
-    if len(s_bytes) > 0 and s_bytes[0] == ord("-"):
+    if len(s_bytes) > 0 and s_bytes[0] == UInt8(ord("-")):
         sign = -1
         i = 1
 
     # Parse digits
     while i < len(s_bytes):
-        if s_bytes[i] >= ord("0") and s_bytes[i] <= ord("9"):
+        if s_bytes[i] >= UInt8(ord("0")) and s_bytes[i] <= UInt8(ord("9")):
             result = result * 10 + (Int(s_bytes[i]) - ord("0"))
         i += 1
